Config and Script Updates

+++ scripts/monitor.sh
@@ -22,13 +22,15 @@ while true; do
     succ=$(printf "%i" $(grep '| common |' /home/ubuntu/hll3_opennebula/caliper/caliper-logs.txt | awk '{print $4}' | tail -n 1))
     echo "Caliper Failure Status - diff: {$diff}, succ: {$succ}"
     if [ $diff == 0 ] || [ $succ == 0 ]; then
-        echo "Caliper Failed"
-        failcount=$((failcount+1))
-        echo "${failcount} Failed - Caliper Restart - Failure Status - diff: {$diff}, succ: {$succ}" >> /home/ubuntu/hll3_opennebula/caliper/failure_logs.txt
-        pkill -9 -f ./scripts/caliper_run.sh
-        pkill -9 -f caliper-manager
-        pkill -9 -f caliper-logs.txt
-        sleep 360s
+        if [ ! -f /home/ubuntu/hll3_opennebula/check.txt ]; then
+            echo "Caliper Failed"
+            failcount=$((failcount+1))
+            echo "${failcount} Failed - Fabric & Caliper Restart - Failure Status - diff: {$diff}, succ: {$succ}" >> /home/ubuntu/hll3_opennebula/caliper/failure_logs.txt
+            pkill -9 -f ./scripts/caliper_run.sh
+            pkill -9 -f caliper-manager
+            pkill -9 -f caliper-logs.txt
+            sleep 420s
+        fi
     fi
     #sed -i 1,200d /home/ubuntu/hll3_opennebula/caliper/caliper-logs.txt
     #if [ $count == 1 ]; then
